Rename modal in module
