+++ front/app/modules/commercial/user_custom_fields/index.tsx
@@ -15,7 +15,7 @@ import {
   IUsersByDomicile,
   IUsersByRegistrationField,
 } from './services/stats';
-import SignUpInModal from './citizen/components/SignUpInModal';
+import UserCustomFieldsSignUpInModal from './citizen/components/UserCustomFieldsSignUpInModal';
 
 declare module 'resources/GetSerieFromStream' {
   export interface ISupportedDataTypeMap {
@@ -127,7 +127,9 @@ const configuration: ModuleConfiguration = {
     ),
     'app.containers.Admin.settings.registration': AllCustomFields,
     'app.containers.Admin.settings.registrationHelperText': RegistrationQuestions,
-    'app.containers.App.signUpInModal': (props) => <SignUpInModal {...props} />,
+    'app.containers.App.signUpInModal': (props) => (
+      <UserCustomFieldsSignUpInModal {...props} />
+    ),
   },
 };
